Add class Grid to contain related info and methods
Declare and define class Grid in grid.h and grid.cpp. Also copy
definition of struct Cell into the beginning of grid.h because it is
required by class Grid and their concepts are closely related (to
represent the playing grid and cell).

With class Grid, the program will be able to play with different sizes
of grids at runtime, due to the dynamic 2-D array stored by the class.
Class Grid also stores information about number of rows and columns.
This eases the need to pass three arguments to another function at the
same time to inform about a grid (grid_array, num_row, num_col).

All the previous functions working directly with the grid and closely
related to the grid is incorporated into class Grid as methods for
convenient usage.