Apply class Grid into game implementation
Remove all the old functions that only work on a 5x5 grid and that must
be provided with three arguments (grid[][5], row, col), and replace them
with the use of class Grid and its methods.

Now the game asks for user to input number of rows and columns to
generate a grid of the required size.

Also remove temporary make rules and C++ codes used for testing the game
implementation. Remove temporary executable from gitignore.