Add NewGameMenu() and its unit test
It will prompt prompt user to input the dimension of the grid to play
with. Then it return a Dimension struct that contains this info.

Make PrintScreen() accept an ostream object for better testability of
functions that use it. (NewGameMenu() being one)

Increase maximum characters limit of istream::ignore from 100 to 256